[win32] Fix pointer to integer casts
'long' will be 32b on Windows, even on x64 targets. For this reason,
we can't use "long" when casting a pointer.

We'll use uintptr_t instead.

+++ src/include/err.h
@@ -18,12 +18,12 @@ static inline void *ERR_PTR(long error)
 
 static inline long PTR_ERR(const void *ptr)
 {
-  return (long) ptr;
+  return (uintptr_t) ptr;
 }
 
 static inline long IS_ERR(const void *ptr)
 {
-  return IS_ERR_VALUE((unsigned long)ptr);
+  return IS_ERR_VALUE((uintptr_t)ptr);
 }
 
 #endif
